requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
requests==2.31.0
python-dotenv==1.0.0
geopandas==0.14.1
folium==0.14.0
ultralytics==8.0.228
